requirements.txt
asgiref==3.8.1
attrs==23.2.0
build==1.2.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
cryptography==42.0.8
dj-database-url==2.2.0
Django==5.0.6
django-anymail==10.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-filter==24.2
django-ratelimit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
Faker==25.9.1
flake8==7.1.0
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
install==1.3.5
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
mccabe==0.7.0
openpyxl==3.1.4
packaging==24.1
pillow==10.3.0
pip-review==1.3.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
treblle==0.14
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
